+++ services/one-app/src/app/(site)/login/_component/SocialLogins.tsx
@@ -2,45 +2,36 @@
 
 import React from 'react';
 import { useRouter } from 'next/navigation';
-import { useQuery } from '@tanstack/react-query';
 
-import { SocialSignInType } from '@/model/Auth';
+import { SocialLoginButton } from './SocialLoginButton';
 import { getRedirectUrl } from '../_lib/getRedirectUrl';
 import { socialLoginOptions } from '../_lib/socialLoginOptions';
-import { SocialLoginButton } from './SocialLoginButton';
+import { SocialSignInType } from '@/model/Auth';
+import { APIResponseCode } from '@/common/constants/api';
 
 export const SocialLogins: React.FC = () => {
-  const createSocialLoginQuery = (socialType: SocialSignInType) =>
-    useQuery({
-      enabled: false,
-      staleTime: Infinity,
-      queryKey: ['oauth', socialType],
-      select: (res) => res.result.redirectUrl,
-      queryFn: () => getRedirectUrl(socialType),
-    });
-
-  const socialQueries = Object.values(SocialSignInType).reduce(
-    (acc, socialType) => {
-      acc[socialType] = createSocialLoginQuery(socialType);
-      return acc;
-    },
-    {} as Record<SocialSignInType, ReturnType<typeof createSocialLoginQuery>>,
-  );
-
   const router = useRouter();
+
   const handleLogin = async (socialType: SocialSignInType) => {
-    const { data: redirectUrl, isError } =
-      await socialQueries[socialType].refetch();
+    try {
+      const { code, result } = await getRedirectUrl(socialType);
 
-    if (isError || !redirectUrl) {
-      alert('로그인 정보를 불러오는데 실패했어요.');
-      return;
+      if (code !== APIResponseCode.SUCCESS) {
+        // alert(RESPONSE_MESSAGES[code]);
+        alert('로그인 정보를 불러오는데 실패했어요. (서버)');
+        return;
+      }
+      router.push(result.redirectUrl);
+    } catch (error) {
+      alert('로그인 정보를 불러오는데 실패했어요. (unknown)');
     }
-
-    router.push(redirectUrl);
   };
 
   return socialLoginOptions.map((option) => (
-    <SocialLoginButton key={option.social} {...option} onLogin={handleLogin} />
+    <SocialLoginButton
+      key={option.social}
+      {...option}
+      onLogin={() => handleLogin(option.providerType)}
+    />
   ));
 };
