+++ services/one-app/src/app/(site)/login/page.tsx
@@ -8,7 +8,7 @@ export const metadata: Metadata = {
   description: 'Generated by create next app',
 };
 
-export default async function Login() {
+export default function Login() {
   return (
     <main className="relative min-h-screen overflow-hidden bg-black">
       <HelloOnLogin />
