+++ services/one-app/src/app/_components/MSWComponent.tsx
@@ -1,4 +1,5 @@
 'use client';
+
 import { useEffect } from 'react';
 
 export const MSWComponent = () => {
