--- services/one-app/src/common/constants/index.ts
@@ -1,4 +1,0 @@
-export enum CookieKey {
-  ACCESS_TOKEN = 'AHAHCHUL_ACCESS_TOKEN',
-  REFRESH_TOKEN = 'AHAHCHUL_REFRESH_TOKEN',
-}
